formatted code, documented all but WriteFiles function. Only commented WriteFiles a little

+++ Tokenizer.cpp
@@ -36,25 +36,30 @@ Both of these actions will exclude single quote characters that appear
 at the begining or end of a word, but any single quotes within the
 word will still be considered valid.
 ***********************************************************************/
-void tokenize( const string& str, vector<string>& tokens, const string& valid = " " )
+void tokenize ( const string& str, vector<string>& tokens, const string& valid = " " )
 {
     // skip delimiters to start of first token
-    int tokenStart = str.find_first_of( valid, 0 );
+    int tokenStart = str.find_first_of ( valid, 0 );
+
     // find next delimiter (i.e., end of first token)
-    int tokenEnd = str.find_first_not_of( valid, tokenStart );
+    int tokenEnd = str.find_first_not_of ( valid, tokenStart );
 
     // loop through input string
-    while ( (unsigned)tokenStart != string::npos )
+    while ( tokenStart != (signed) string::npos )
     {
+        // only emebeded single quotes are valid
         if ( str[tokenStart] == '\'')
             tokenStart++;
         if ( str[tokenEnd] == '\'')
             tokenEnd--;
+
         // found a token, add it to the vector
-        tokens.push_back( str.substr( tokenStart, tokenEnd - tokenStart ) );
+        tokens.push_back( str.substr ( tokenStart, tokenEnd - tokenStart ) );
+
         // skip delimiters to start of next token
-        tokenStart = str.find_first_of( valid, tokenEnd );
+        tokenStart = str.find_first_of ( valid, tokenEnd );
+
         // find next delimiter (end of token)
-        tokenEnd = str.find_first_not_of( valid, tokenStart );
+        tokenEnd = str.find_first_not_of ( valid, tokenStart );
     }
 }
